+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/cache/DriverSupportCacheImpl.java
@@ -3,54 +3,96 @@ package com.dwarfeng.judge.impl.cache;
 import com.dwarfeng.judge.sdk.bean.entity.FastJsonDriverSupport;
 import com.dwarfeng.judge.stack.bean.entity.DriverSupport;
 import com.dwarfeng.judge.stack.cache.DriverSupportCache;
-import com.dwarfeng.subgrade.impl.cache.RedisBaseCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Repository
 public class DriverSupportCacheImpl implements DriverSupportCache {
 
-    private final RedisBaseCache<StringIdKey, DriverSupport, FastJsonDriverSupport> baseCache;
+    private final RedisBatchBaseCache<StringIdKey, DriverSupport, FastJsonDriverSupport>
+            driverSupportBatchBaseDelegate;
 
-    public DriverSupportCacheImpl(RedisBaseCache<StringIdKey, DriverSupport, FastJsonDriverSupport> baseCache) {
-        this.baseCache = baseCache;
+    public DriverSupportCacheImpl(
+            RedisBatchBaseCache<StringIdKey, DriverSupport, FastJsonDriverSupport> driverSupportBatchBaseDelegate
+    ) {
+        this.driverSupportBatchBaseDelegate = driverSupportBatchBaseDelegate;
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return baseCache.exists(key);
+        return driverSupportBatchBaseDelegate.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public DriverSupport get(StringIdKey key) throws CacheException {
-        return baseCache.get(key);
+        return driverSupportBatchBaseDelegate.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(DriverSupport value, long timeout) throws CacheException {
-        baseCache.push(value, timeout);
+        driverSupportBatchBaseDelegate.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        baseCache.delete(key);
+        driverSupportBatchBaseDelegate.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        baseCache.clear();
+        driverSupportBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return driverSupportBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return driverSupportBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<DriverSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return driverSupportBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<DriverSupport> entities, long timeout) throws CacheException {
+        driverSupportBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        driverSupportBatchBaseDelegate.batchDelete(keys);
     }
 }