+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/cache/JudgerSupportCacheImpl.java
@@ -3,54 +3,96 @@ package com.dwarfeng.judge.impl.cache;
 import com.dwarfeng.judge.sdk.bean.entity.FastJsonJudgerSupport;
 import com.dwarfeng.judge.stack.bean.entity.JudgerSupport;
 import com.dwarfeng.judge.stack.cache.JudgerSupportCache;
-import com.dwarfeng.subgrade.impl.cache.RedisBaseCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Repository
 public class JudgerSupportCacheImpl implements JudgerSupportCache {
 
-    private final RedisBaseCache<StringIdKey, JudgerSupport, FastJsonJudgerSupport> baseCache;
+    private final RedisBatchBaseCache<StringIdKey, JudgerSupport, FastJsonJudgerSupport>
+            judgerSupportBatchBaseDelegate;
 
-    public JudgerSupportCacheImpl(RedisBaseCache<StringIdKey, JudgerSupport, FastJsonJudgerSupport> baseCache) {
-        this.baseCache = baseCache;
+    public JudgerSupportCacheImpl(
+            RedisBatchBaseCache<StringIdKey, JudgerSupport, FastJsonJudgerSupport> judgerSupportBatchBaseDelegate
+    ) {
+        this.judgerSupportBatchBaseDelegate = judgerSupportBatchBaseDelegate;
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws CacheException {
-        return baseCache.exists(key);
+        return judgerSupportBatchBaseDelegate.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public JudgerSupport get(StringIdKey key) throws CacheException {
-        return baseCache.get(key);
+        return judgerSupportBatchBaseDelegate.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void push(JudgerSupport value, long timeout) throws CacheException {
-        baseCache.push(value, timeout);
+        judgerSupportBatchBaseDelegate.push(value, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws CacheException {
-        baseCache.delete(key);
+        judgerSupportBatchBaseDelegate.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void clear() throws CacheException {
-        baseCache.clear();
+        judgerSupportBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return judgerSupportBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return judgerSupportBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<JudgerSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return judgerSupportBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<JudgerSupport> entities, long timeout) throws CacheException {
+        judgerSupportBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        judgerSupportBatchBaseDelegate.batchDelete(keys);
     }
 }