+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/cache/DriverSupportCache.java
@@ -2,7 +2,7 @@ package com.dwarfeng.judge.stack.cache;
 
 import com.dwarfeng.judge.stack.bean.entity.DriverSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.cache.BaseCache;
+import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 
 /**
  * 驱动器支持缓存。
@@ -10,5 +10,5 @@ import com.dwarfeng.subgrade.stack.cache.BaseCache;
  * @author DwArFeng
  * @since beta-1.1.0
  */
-public interface DriverSupportCache extends BaseCache<StringIdKey, DriverSupport> {
+public interface DriverSupportCache extends BatchBaseCache<StringIdKey, DriverSupport> {
 }