+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/dao/JudgerSupportDao.java
@@ -2,7 +2,7 @@ package com.dwarfeng.judge.stack.dao;
 
 import com.dwarfeng.judge.stack.bean.entity.JudgerSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.dao.BaseDao;
+import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
 import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
 import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
 
@@ -12,6 +12,6 @@ import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
  * @author DwArFeng
  * @since beta-1.1.0
  */
-public interface JudgerSupportDao extends BaseDao<StringIdKey, JudgerSupport>, EntireLookupDao<JudgerSupport>,
+public interface JudgerSupportDao extends BatchBaseDao<StringIdKey, JudgerSupport>, EntireLookupDao<JudgerSupport>,
         PresetLookupDao<JudgerSupport> {
 }