+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/service/JudgerSupportMaintainService.java
@@ -3,7 +3,7 @@ package com.dwarfeng.judge.stack.service;
 import com.dwarfeng.judge.stack.bean.entity.JudgerSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
-import com.dwarfeng.subgrade.stack.service.CrudService;
+import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
 
@@ -13,7 +13,7 @@ import com.dwarfeng.subgrade.stack.service.PresetLookupService;
  * @author DwArFeng
  * @since beta-1.1.0
  */
-public interface JudgerSupportMaintainService extends CrudService<StringIdKey, JudgerSupport>,
+public interface JudgerSupportMaintainService extends BatchCrudService<StringIdKey, JudgerSupport>,
         EntireLookupService<JudgerSupport>, PresetLookupService<JudgerSupport> {
 
     String ID_LIKE = "id_like";